Add files via upload
This is a story based Treasure hunt game. The game is dependent on the user's action. if he makes the right decision he will go to next level ,else the game ends. nested if-else statements were used to built the logic of this game and it could be even further developed to continue the story.